[Runtime] Runtime e2e cpu test.
